contacts controller code refactoring
