+++ controller/contacts/index.js
@@ -1,0 +1,17 @@
+const {controllerWrapper} = require("../../helpers");
+
+const getAllContacts = require("./getAllContacts");
+const getContactById = require("./getContactById");
+const addContact = require("./addContact");
+const removeContact = require("./removeContact");
+const updateContact = require("./updateContact");
+const updateStatusContact = require("./updateStatusContact");
+
+module.exports = {
+  getAllContacts: controllerWrapper(getAllContacts),
+  getContactById: controllerWrapper(getContactById),
+  addContact: controllerWrapper(addContact),
+  removeContact: controllerWrapper(removeContact),
+  updateContact: controllerWrapper(updateContact),
+  updateStatusContact: controllerWrapper(updateStatusContact),
+};
